Sliding View Garbage Collector
The Sliding View GC algorithm was created in pursuance of a masters
thesis around 1999 by Yossi Levanoni. This repository contains the
original code and the PDF file of the thesis. The code was integrated
into Sun's interpreter-based JVM of that time, and therefore may contain
residual Sun IP. Similarly, since the code was created for a thesis at
the Technion - Israel's Institute of Technology, the Technion may have
rights to the IP as well. The author (Yossi Levanoni) assumes no
responsibility for usage incompatible with possible terms imposed by Sun
(now Oracle) and/or the Technion. The author also assume no liability
from any other unexpected consequences of reading or reusing this code.